requirements.txt
astroid==2.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
clarifai==2.6.2
configparser==3.8.1
cryptography==2.9.2
decorator==4.4.2
emoji==0.6.0
future==0.18.2
googleapis-common-protos==1.52.0
grpcio==1.31.0
idna==2.10
isort==4.3.21
jsonschema==2.6.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
protobuf==3.13.0
pycparser==2.20
pylint==2.5.3
PySocks==1.7.1
python-telegram-bot==12.7
requests==2.24.0
six==1.15.0
toml==0.10.1
tornado==6.0.4
typed-ast==1.4.1
urllib3==1.25.10
wrapt==1.12.1
